move normalize and denormalize into perturb  (#60)
This PR move the normalize and denormalize into perturb as a first step to decouple the normalization (which is model dependent) from the time loop. To do this requires providing perturb with the means and stds of the channels for all models. To that I added two numpy files to channel_means.py and channel_stds.py which are dicts that contain the largest dataset we currently have (235 channels).