Run graphcast on the correct device (#153)
* Fix device-emplacement error in graphcast

When multiple GPUs are present graphcast was not running on the
specified device.  Jax places the model on the device of the input data
it is passed. However, when converting the torch tensor to an
xarray.Dataset of jax arrays the data were being cast to CPU numpy
arrays. Using xarray_jax.Variable resolves this.

* improve name of _get_array_module